My first paid work
This was my first paid freelance work from a year ago, and until then I
had never done any html, css, or javascript. It was definitely a
challenge learning all this on my own, and this project took me 2
months to finish. Big thanks to Suresh for giving me a chance!